Scaling: Add 'Distance Scale' Option

+++ src/Enemies/Enemy.c
@@ -37,7 +37,7 @@ long		gNumEnemies;
 
 void InitEnemyManager(void)
 {
-long	i;
+short	i;
 
 	gTheQueen = nil;
 	gAntKingObj = nil;
